enforce clean architecture: service interfaces, DTO validation constraints, PageResult boundary, remove dead swagger code

+++ backend/src/main/java/at/technikumwien/mse25/awt/hotelmario/components/rooms/api/dtos/v1/RoomDto.java
@@ -1,6 +1,7 @@
 package at.technikumwien.mse25.awt.hotelmario.components.rooms.api.dtos.v1;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.Builder;
 import lombok.Value;
@@ -13,6 +14,6 @@ public class RoomDto {
     @JsonProperty("title") String title;
     @JsonProperty("description") String description;
     @JsonProperty("imageUrl") String imageUrl;
-    @JsonProperty("pricePerNight") Double pricePerNight;
+    @JsonProperty("pricePerNight") BigDecimal pricePerNight;
     @JsonProperty("extras") List<ExtraDto> extras;
 }
